+++ projects/iucn/src/app/+regional/regional/regional.component.ts
@@ -1,8 +1,8 @@
-import {Component, OnInit, ChangeDetectionStrategy, OnDestroy} from '@angular/core';
-import {Observable, of as ObservableOf, Subscription} from 'rxjs';
-import {map, startWith, switchMap} from 'rxjs/operators';
+import { Component, OnInit, ChangeDetectionStrategy, OnDestroy } from '@angular/core';
+import { Observable, of as ObservableOf, Subscription } from 'rxjs';
+import { map, startWith, switchMap } from 'rxjs/operators';
 import { TranslateService } from '@ngx-translate/core';
-import {ActivatedRoute, Router} from '@angular/router';
+import { ActivatedRoute, Router } from '@angular/router';
 import {RegionalFilterQuery, RegionalService, REGIONAL_DEFAULT_YEAR} from '../../iucn-shared/service/regional.service';
 
 export type RegionalListType = 'status'|'species';
